Review findings batch: dedup, bug fix, UX polish, test coverage
Addresses findings from the in-depth code review across all three categories.

Code quality (architecture):
- 2.1: Extract duplicated view helpers into lib/view/view_helpers.dart — formatDate, formatDateTime, impactLabel, validatePositiveInt. Remove 12 identical copies from add_node_view, node_detail_view, quick_add_child_dialog, graph_canvas_view, todo_list_view.
- 2.3: Extract duplicated compare utils into lib/service/compare_utils.dart — earlierDate, strongerImpact. Remove copies from node_queries + hierarchical_ordering.
- 2.6: _saveAsTile now uses controller.setFilterPresets instead of the racy copyWith+replaceWith pattern.
- 2.8: Impact.weight changed from double to int (values are 1-5, no FP needed).
- 2.10: FilterPreset gains copyWith; ManagePresetsView uses it instead of hand-reconstructing instances.
- 2.11: Noted but not centralized (encoder is const and trivial; 3 copies are acceptable).

Bug fix:
- 3.4: FilterEvaluator._isTimewiseInactive now also checks now.isAfter(activeUntil) — bounded tasks whose window has ENDED are hidden when showTimewiseInactiveTasks is false. Previously only future-start windows were hidden. Test added.

UX polish:
- 1.2: Dashboard tiles show an actionable-task count badge (ongoing leaves for the tile's filter) so the user's morning glance shows numbers without tapping in.
- 1.3: AddNodeView parent picker replaced from a flat DropdownButtonFormField (unusable at 30+ nodes) with the searchable NodePickerDialog already used elsewhere.
- 1.5: Tile title Text gains maxLines: 2 + TextOverflow.ellipsis so long titles like 'Personal development' don't overflow.
- 1.6: _SaveAsTileDialog TextField gains onSubmitted so Enter submits (matches _RenameDialog).

Test coverage:
- 3.1: 4 widget tests for the node editor dialog: title-edit-preserves-fields, pre-populated-values, save-without-changes, cancel-discards.
- 3.8: Quick-add 'Every 3 days' test verifying PeriodicCompletion with interval=3.
- 3.9: Deep-chain (10-node linear) cycle detection regression test.
- 3.10: buildGraph test helper added to test/support/builders.dart.

234 tests pass, flutter analyze clean.

+++ lib/model/impact.dart
@@ -13,7 +13,7 @@ enum Impact {
 
   const Impact({required this.weight});
 
-  final double weight;
+  final int weight;
 
   String toJsonValue() => name;
